requirements.txt
flask>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
wtforms>=3.0.0
flask-wtf>=1.0.0
requests>=2.25.0